move unit tests into folder
